Updated Job Metrics and implemented where they should be updated through out(Currently updating all the job metrics directly to the PCB, but have the other way commented out). Only kept the avg wait time and avg completion time in Metrics.h. % of RAM and %of Cache per process is asigned when PCB filled in loader.
Made CPUMetrics similar to PCB and keep track of it via map in the computer.cpp.  Now have a m_CPUMetrics in the CPU class which is assigned in the short term scheduler. Currently have the Read and Write key as static but would like to have a specific write key for each CPU held by the CPUMetrics(which this is currently implemented but commented out).